curve out grep filter to independent package
