+++ filter/grep/grep.go
@@ -1,4 +1,4 @@
-package filter
+package grep
 
 import (
 	"bytes"
@@ -6,6 +6,7 @@ import (
 	"io"
 	"regexp"
 
+	"github.com/koron/nvgd/filter"
 	"github.com/koron/nvgd/internal/filterbase"
 	"github.com/koron/nvgd/resource"
 )
@@ -18,7 +19,7 @@ type Grep struct {
 
 	re    *regexp.Regexp
 	match bool
-	lf    LineFilter
+	lf    filter.LineFilter
 	lnum  bool
 	cnum  int
 
@@ -27,12 +28,12 @@ type Grep struct {
 }
 
 // NewGrep creates an instance of grep filter.
-func NewGrep(r io.ReadCloser, re *regexp.Regexp, match bool, lf LineFilter, lnum bool, cnum int) *Grep {
+func NewGrep(r io.ReadCloser, re *regexp.Regexp, match bool, lf filter.LineFilter, lnum bool, cnum int) *Grep {
 	g := &Grep{
 		reader: filterbase.NewLineReader(r),
 		re:     re,
 		match:  match,
-		lf:     TrimEOL.Chain(lf),
+		lf:     filter.TrimEOL.Chain(lf),
 		lnum:   lnum,
 		cnum:   cnum,
 	}
@@ -90,7 +91,7 @@ func (g *Grep) output(buf *bytes.Buffer, lnum int, data []byte) error {
 	return err
 }
 
-func newGrep(r *resource.Resource, p Params) (*resource.Resource, error) {
+func newGrep(r *resource.Resource, p filter.Params) (*resource.Resource, error) {
 	re, err := regexp.Compile(p.String("re", ""))
 	if err != nil {
 		return nil, err
@@ -98,18 +99,18 @@ func newGrep(r *resource.Resource, p Params) (*resource.Resource, error) {
 	match := p.Bool("match", true)
 	lnum := p.Bool("number", false)
 	cnum := p.Int("context", 0)
-	var lf LineFilter
+	var lf filter.LineFilter
 	// field filter
 	var (
 		field = p.Int("field", 0)
 		delim = []byte(p.String("delim", "\t"))
 	)
 	if field > 0 {
-		lf = lf.Chain(NewCutLF(delim, field-1))
+		lf = lf.Chain(filter.NewCutLF(delim, field-1))
 	}
 	return r.Wrap(NewGrep(r, re, match, lf, lnum, cnum)), nil
 }
 
 func init() {
-	MustRegister("grep", newGrep)
+	filter.MustRegister("grep", newGrep)
 }
